move validate rpc with chain id to util
